update merge request content for http://ess-repos01.wisers.com:9980/Sentiment/API/canton-target-sentiment/merge_requests/4

+++ src/dim_reduction.py
@@ -21,7 +21,7 @@ def get_local_emb(local_path: str):
             split_result = line.rstrip().split(" ")
             vocabs.append(split_result[0].replace('\u2028', ''))
             vectors.append(split_result[1:])
-    vectors = np.array(vectors, dtype=float)
+    vectors = np.array(vectors, dtype=np.float32)
     vocabs = np.array([vocabs])
     logger.info("  Embeddings size = '%s'", str(vectors.shape)) 
     return vocabs, vectors
